Reorganize the "wind" related stages
It's likely that these stages will be merged or removed, but it gives me
a target. If I can create enough content for these sections to exist,
and possibly stand on their own, then they'll stay, but they're probably
mental placeholders as I figure out how to structure this thing.