Merge HTML and latex index files
Hope I didn't break anything. Until this commit the project was using a
quirky design that split the HTML and latex index files, and I know
there was a reason when I first did it, but now it just adds
(seemingly?) unnecessary clutter, causes the builders to complain about
nested numbered toctrees (since it was parsing both `index.rst` and
`index-latex.rst`), and added duplications (which are best avoided).

One concern is that this commit removes an `:orphan:` directive in the
index file. Apparently the latex builder would skip a title without it,
but that doesn't appear to be the case. I don't remember how it came
about, but keep an eye out if it breaks chapter headings.