Reset font to Latin Modern 11pt
I was using Tex Gyre Heros as a Helvetica substitute, but the sans-serif
was really bothering me. I tried Tex Gyre Termes as a Times New Roman
replacement, but then I questioned why I was bothering in the first
place? If CalPoly complains I'll switch to Termes, otherwise I'll stick
to the default: it looks great.