Replace introduction with stub
The old introduction was centered around flight reconstruction, but the
entire paper is focus on the parametric model. It's confusing to see all
that discussion of wind fields, statistical filtering, etc, then never
discuss it in the paper.

About to send a draft Slivovsky; the abstract summarizes the paper
already, I want her to focus on the "important" content. I'll figure out
an introduction later.